Move exceptions to message_layer
