--- src/main/java/com/streamr/client/exceptions/MalformedMessageException.java
@@ -1,13 +1,0 @@
-package com.streamr.client.exceptions;
-
-public class MalformedMessageException extends RuntimeException {
-
-    public MalformedMessageException(String message) {
-        super(message);
-    }
-
-    public MalformedMessageException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-}
