+++ src/main/java/com/streamr/client/protocol/message_layer/AbstractGroupKeyMessage.java
@@ -1,6 +1,5 @@
 package com.streamr.client.protocol.message_layer;
 
-import com.streamr.client.exceptions.MalformedMessageException;
 import com.streamr.client.protocol.message_layer.StreamMessage.MessageType;
 import com.streamr.client.utils.ValidationUtil;
 import java.io.IOException;
