+++ src/main/java/com/streamr/client/protocol/message_layer/MalformedMessageException.java
@@ -1,0 +1,11 @@
+package com.streamr.client.protocol.message_layer;
+
+public class MalformedMessageException extends RuntimeException {
+  public MalformedMessageException(String message) {
+    super(message);
+  }
+
+  public MalformedMessageException(String message, Throwable cause) {
+    super(message, cause);
+  }
+}
