+++ src/main/java/com/streamr/client/protocol/message_layer/StreamMessageAdapter.java
@@ -3,7 +3,6 @@ package com.streamr.client.protocol.message_layer;
 import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.JsonReader;
 import com.squareup.moshi.JsonWriter;
-import com.streamr.client.exceptions.MalformedMessageException;
 import com.streamr.client.exceptions.UnsupportedMessageException;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
