+++ src/main/java/com/streamr/client/utils/MessageCreationUtil.java
@@ -3,7 +3,7 @@ package com.streamr.client.utils;
 import com.streamr.client.exceptions.InvalidGroupKeyException;
 import com.streamr.client.exceptions.InvalidGroupKeyRequestException;
 import com.streamr.client.exceptions.InvalidGroupKeyResponseException;
-import com.streamr.client.exceptions.MalformedMessageException;
+import com.streamr.client.protocol.message_layer.MalformedMessageException;
 import com.streamr.client.exceptions.SigningRequiredException;
 import com.streamr.client.protocol.message_layer.GroupKeyAnnounce;
 import com.streamr.client.protocol.message_layer.GroupKeyErrorResponse;
